Separate file and bundle tests. (#100)
